temp - WIP - IocContainerScope - disposal of wrapped IServiceProviders 2
Made sure wrapped scopes get disposed, even if the stored IServiceProvider does not implement IDisposable or IAsyncDisposable.

And fixed hot path GetServiceFunc assignment for non-root Scopes; on scope creation it has not been assigned to parent's GetServiceFunc, now added/fixed.

Finally, ensured on disposal of non-root Scope that it only disposes non-root wrapped IServiceProviders / wrapped scopes.
When scope is created, but a wrapped IServiceProvider was unable to create a scope, it continues using it as root scope in non-root scope. Disposing it on scope disposal would make problems for the real root scope, so now it has proper checks; it simply checks if IContainerScopeFactory is registered or not, if not no scope is creatable from that wrapped Provider(s).